Add all prompts to multiplication, BUT!
There are some bugs appear after adding the prompts, but I can't finish
it now. To avoid affecting add/subtraction, I put some codes in
annotation. If you continue to do it, you could reuse the codes in
animator_graph with the annotation "Adding prompts in multiplication
hasn't been done. If you continue to do it, the codes following could be
reused."